chap10/instagram/backend/backend/requirements/common.txt
django ~= 3.0.0
djangorestframework
django-cors-headers
pillow

django-pydenticon